requirements.txt
attrs==20.3.0
branca==0.4.1
brotlipy==0.7.0
Cartopy==0.18.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
et-xmlfile==1.0.1
ExifRead==2.1.2
Fiona==1.8.18
folium==0.11.0
GDAL==3.1.4
geographiclib==1.50
geopandas==0.8.1
geopy==2.1.0
gpsphoto==2.2.3
idna==2.10
jdcal==1.4.1
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
mapclassify==2.4.2
MarkupSafe==1.1.1
matplotlib==3.3.3
munch==2.5.0
networkx==2.5
numpy==1.19.4
olefile==0.46
openpyxl==3.0.5
pandas==1.2.0
piexif==1.1.3
Pillow==8.0.1
pip==20.3.3
pycparser==2.20
pygeos==0.8
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyproj==2.6.1.post1
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyshp==2.1.2
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
Rtree==0.9.7
scikit-learn==0.24.0
scipy==1.5.3
setuptools==49.6.0.post20201009
Shapely==1.7.1
six==1.15.0
threadpoolctl==2.1.0
tornado==6.1
urllib3==1.26.2
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
